Exercise 6 - Carousels in sync
Created a "carousel.isml" file for rendering the data on storefront.
Created a controller "Carousel.js" with "Show" endpoint in it.
In this controller created a function to get random images from a link and added loop to it.
Created "carousel.js" for adding functionality of the slider and included it in "main.js".
Created "carousel.scss" for styling the page and Imported it in "global.scss" file.